+++ src/lib/pages/instantiate/instantiate.tsx
@@ -33,7 +33,7 @@ import {
 } from "lib/services/amplitude";
 import { getCodeIdInfo } from "lib/services/code";
 import type { HumanAddr, Token, U } from "lib/types";
-import { MsgType } from "lib/types";
+import { InstantiatePermission, MsgType } from "lib/types";
 import {
   composeMsg,
   demicrofy,
@@ -142,7 +142,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
         const permission = data.code_info.instantiate_permission;
         if (
           address &&
-          (permission.permission === "Everybody" ||
+          (permission.permission === InstantiatePermission.EVERYBODY ||
             permission.addresses.includes(address) ||
             permission.address === address)
         )